Added dynamic URLs in flask
I needed to route the user to the team page with their chosen team id. This allows for specific team information (e.g messages, events) to be loaded with the same logic and html layout. This documentation supported me 'https://www.geeksforgeeks.org/generating-dynamic-urls-in-flask/%22'. I will use this for future messaging and event pages for specific teams.