Player can join a team + coach can create a team
Added a searchTeam page so users can join a team using a search bar. Currently the user can only join/be in one team. Later on for scalbility, this can be expanded upon to allow the user to join multiple teams, but for now it will just be one team. Also added a createTeam page with a FlaskForm that can be only accessed by coaches to create a new team. It's all functioning as intended but some of the code can be improved to be more maintainable.