player can do attendance + coach can see attendance
Added a modal for player to rsvp to team events. to do this created a event attendance table which had references to event & user id from other tables and also a check constraint of 'status'. had some difficulties with the SQL language but had assistance from w3schools and copilot. also added so the coach can see the number of people attending/not attending through a badge.