Sprint 2.0 completed
Completed sprint 2 review. Added an attendance summary for coaches to shows the number of attendees and user responses. Had to implement back-end logic in my dbHandler and added a modal to pop up for coaches to see summary. Added checks to see if the user is part of the team before redirecting to a team page to improve security and privacy.