responsive interface + CSS improvements
I worked on improving the CSS for my whole PWA to improve the looks and make it look like a final product. Due to CSS being so expansive, I tried not to delve to much into it. I used GitHub Copilot to simplify and streamline my workflow while working on this. Whilst testing my CSS, I came into an issue with authentication. During 2FA, the user could access the team navbar by redirecting themselves to a differnet page as the session was set up once logging in. To correct this, I created a temporary session for the user while doing 2FA and then set the actual session once the 2FA was confirmed. This prevented the threat of broken authentication. I also improved upon the dark mode I implemented in the previous commit. Furthermore, I added a large font option to address any accessbility needs. While testing these options,  I ran into another problem with my CSP not loading the script to set the theme and large font if needed. There was a mistake in my csp_header decorator that was fixed.