can join multiple teams + team profile pics
Previously the user could only join one team as my database structure limited this but from  client feedback I decided to implement a way for users to join multiple teams. I had to create another table in my database to create a many to many relationship between the the secure_user_9f table and the new users_teams table. I used this website to clarify the relationships in SQL for me 'https://www.geeksforgeeks.org/sql/relationships-in-sql-one-to-one-one-to-many-many-to-many/'.  This many to many relationship allowed  me to store multiple team_ids for a single user. I had to retweak some of my dbHandler functionst to accommodate for this change. I also added the ability to leave a team to support this new functionality. Futhermore from more client feedback, I added the abiltiy for coaches to upload a profile picture when creating a team so this can be displayed on the home page for that specific team. In the backend, the file is sanitised and verified and then saved securedly.