Sprint 4.0 completed
General tweaks and debugging were done here to finish off this sprint. This included a fix when submitting the create team event form. The recurrence_end attribute of the FlaskForm was intitally required to submit the form but this prevented team events without recurrence to not submit. This was an easy fix to just make the attribute optional. Also updated READme.md.